src/run-tests.sh: Removed "Printcap" line from cups-files.conf
After a recent security fix CUPS is more strict with configuration file syntax and does not start when a directive is missing a required argument.

The line only containing "Printcap" in cups-files.conf prevents cupsd from starting therefore and makes our test failing.

Therefore we have removed the line now.

As cupsd fails silently on errors in config files we also run a syntax check on the files before actually starting CUPS, so we can see what is wrong in the config files when cupsd is not starting.

Also fixed the header comment to describe this script correctly.